backend/pyproject.toml
[project]
name = "reputation-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # DB
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",

    # Task Queue
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "flower>=2.0.0",

    # AI — 콘텐츠 생성
    "anthropic>=0.40.0",

    # AI — SoV 측정
    "openai>=1.55.0",
    "httpx>=0.28.0",

    # AI — 이미지 생성
    "google-cloud-aiplatform>=1.70.0",
    "google-cloud-storage>=2.18.0",

    # Retry
    "tenacity>=9.0.0",

    # Report
    "weasyprint>=63.0",
    "jinja2>=3.1.0",

    # Utils
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "arrow>=1.3.0",
    "python-slugify>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
